Create index.html
Elyas and Sufyan
POLYGLOTS
Multilingual Workbook
Arabic, English, French, Serbo-Croatian, Spanish

Parent or Tutor’s Guide
When reading this book with your child, keep the following in mind:
In this story, Elyas and Sufyan speak five languages and are reliving their childhood adventures around the world. Your child can pretend to be them and guess what they are recognizing or rediscovering in the pictures.
This short story encourages your child to:
1. practice and/or explore different languages,
2. use his or her imagination,
3. think critically by asking and answering questions about the
story.
This is also a writing workbook, which means that space is provided so that your child may practice writing by copying every line of text or answering the questions in the blank column (in one language, or each answer in a different language).
Encourage your child to come up with as many answers as possible. You may prompt her by asking leading questions. If you are reading with more than one child, they can compete!